smart-learning/requirements.txt
scikit-learn==0.18.2
mock
nltk
flask
numpy
word2number
scipy
spacy
elasticsearch
joblib
sklearn_crfsuite
pint